Refactor Neuron
