+++ Assets/Scripts/AI.meta
@@ -1,6 +1,7 @@
 fileFormatVersion: 2
-guid: d8ac06deed315784693a8a7dec21d7bd
-AssemblyDefinitionImporter:
+guid: 3c211bca01667244db0285ff7e859f5b
+folderAsset: yes
+DefaultImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
