+++ Assets/Scripts/AI/AI.asmdef
@@ -1,0 +1,3 @@
+{
+	"name": "AI"
+}
