+++ Assets/Scripts/AI/AI.asmdef.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: a67e022a4c759274599c39ef6766b472
+AssemblyDefinitionImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
