--- Assets/Scripts/NeuralNet/NeuralNet.asmdef
@@ -1,3 +1,0 @@
-{
-	"name": "NeuralNet"
-}
