+++ Assets/Tests/EditMode/DarwinTests.cs
@@ -1,0 +1,16 @@
+using FluentAssertions;
+using RansomeCorp.AI.Evolution;
+using System.Collections;
+using System.Collections.Generic;
+using NUnit.Framework;
+using UnityEngine;
+using UnityEngine.TestTools;
+
+public class DarwinTests
+{
+    [Test]
+    public void DarwinTestsSimplePasses()
+    {
+
+    }
+}
