+++ Assets/Tests/EditMode/DarwinTests.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 7782cf40e78424b40b236c1c4b7d174c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
