+++ Assets/Tests/EditMode/EditMode.asmdef
@@ -5,7 +5,7 @@
         "UnityEngine.TestRunner",
         "UnityEditor.TestRunner",
         "BoundfoxStudios.FluentAssertions",
-        "NeuralNet"
+        "AI"
     ],
     "includePlatforms": [
         "Editor"
